Better change detection: use hash instead of mtime.
We used to look at the modification time to detect changes and decide
whether to reindex a file or not. This was fine if the documentation was
updated in place.

However, in our case, we regenerate the documentation from source
files (with Sphinx) from scratch on each commit. This uploads brand new
files with fresh modification times, even if they did not change. All
files being more recent than what we have in the index, Dokang wrongly
reindexes all of them.

We now compute and store an MD5 hash for each file. It is slower than
getting the modification time, but it better handles our use case.