Remove bogus indexation optimization.
I tried to use `commit(optimize=True)` because it looked smart and
innocent. I *think* that an old version of Whoosh documentation actually
suggested to use it. The current (2.7) documentation is a bit less
prescribing:

    Since optimizing rewrites all the information in the index, it can
    be slow on a large index. It’s generally better to rely on Whoosh’s
    merging algorithm than to optimize all the time.

Indeed, on a database of 100 documents, using `optimize` takes around 3
seconds, while not using it (relying on the default settings) only takes
a few milliseconds. The effect on search is not noticeable enough to
warrant such an overhead on the indexation.