Fix bug in document deletion.
All documents with the same paths were removed from the index. Remember
that paths are not unique. What is unique is the set/path combination.

This is a remnant of early versions of Dokang where document sets could
be stored anywhere (and not under a single directory as it is now with
`dokang.uploaded_docs.dir`), which made us store the relative path,
e.g. `index.html` instead of `<doc_set>/index.html`.

Now that all document sets live under a single directory, we may
reconsider the meaning of "path" and make it include the name of the
directory of the document set.